ci: Add rule for Firedancer review on native programs (#2233)
* ci: Add rule for Firedancer review on native programs

The Firedancer team maintains a line-for-line reimplementation of the
native programs, and until native programs are moved to BPF, those
implementations must exactly match their Agave counterparts. In the
past, the Firedancer team has requested to be included in reviews that
touch code in the native programs.

To make it harder to forget, add an automated to any PRs that touch
native programs. I've omitted test crates and and the ZK Token Proof
program, since I believe that won't be included in Firedancer.

* Add stake and vote in SDK

* Update language